requirements.txt
requests
pandas
numpy
scikit-learn
python-dotenv
tqdm
streamlit
